Allow storing larger IDs for kits by expanding a database column
Update the bundle_entity_id column in the kit_config table from INTEGER to BIGINT and add a conditional migration to handle existing data.

+++ backend/app/models/kit_config.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, Numeric
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean, Numeric
 from sqlalchemy.sql import func
 from ..core.database import Base
 
@@ -6,7 +6,7 @@ from ..core.database import Base
 class KitConfig(Base):
     __tablename__ = "kit_config"
 
-    bundle_entity_id = Column(Integer, primary_key=True)
+    bundle_entity_id = Column(BigInteger, primary_key=True)
     id_evento = Column(Integer, nullable=True)
     kit_nome = Column(String(255), nullable=True)
     tipo_kit = Column(String(100), nullable=True)
